add generic type
