+++ src/main/java/de/muenchen/isi/domain/service/search/SearchPreparationService.java
@@ -1,10 +1,7 @@
 package de.muenchen.isi.domain.service.search;
 
 import de.muenchen.isi.domain.exception.EntityNotFoundException;
-import de.muenchen.isi.domain.mapper.SearchDomainMapper;
-import de.muenchen.isi.domain.model.BaseEntityModel;
 import de.muenchen.isi.domain.model.search.SearchQueryForEntitiesModel;
-import de.muenchen.isi.domain.model.search.SearchResultsModel;
 import de.muenchen.isi.infrastructure.entity.BaseEntity;
 import de.muenchen.isi.infrastructure.entity.Bauvorhaben;
 import de.muenchen.isi.infrastructure.entity.Infrastrukturabfrage;
@@ -18,20 +15,16 @@ import java.lang.annotation.Annotation;
 import java.lang.reflect.Field;
 import java.lang.reflect.ParameterizedType;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import javax.persistence.EntityManager;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.BooleanUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.reflect.FieldUtils;
-import org.hibernate.search.engine.search.common.BooleanOperator;
-import org.hibernate.search.mapper.orm.Search;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.FullTextField;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.GenericField;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.IndexedEmbedded;
@@ -43,7 +36,7 @@ import org.springframework.stereotype.Service;
 @Service
 @RequiredArgsConstructor
 @Slf4j
-public class SucheService {
+public class SearchPreparationService {
 
     private static final Set<Class<? extends Annotation>> SEARCH_INDEX_ANNOTATION = Set.of(
         IndexedEmbedded.class,
@@ -53,61 +46,6 @@ public class SucheService {
         ScaledNumberField.class,
         NonStandardField.class
     );
-
-    private final EntityManager entityManager;
-
-    private final SearchDomainMapper searchDomainMapper;
-
-    public SearchResultsModel searchForEntities(final SearchQueryForEntitiesModel searchQueryInformation)
-        throws EntityNotFoundException {
-        final List<Class<? extends BaseEntity>> searchableEntities = getSearchableEntities(searchQueryInformation);
-        final var searchResults =
-            this.searchForEntities(searchableEntities, searchQueryInformation.getSearchQuery())
-                .map(searchDomainMapper::model2SearchResultModel)
-                .collect(Collectors.toList());
-        final var model = new SearchResultsModel();
-        model.setSearchResults(searchResults);
-        return model;
-    }
-
-    protected Stream<? extends BaseEntityModel> searchForEntities(
-        final List<Class<? extends BaseEntity>> searchableEntities,
-        final String searchQuery
-    ) {
-        final var searchableAttributes = getNamesOfSearchableAttributes(searchableEntities);
-        final var adaptedSearchQuery = this.createAdaptedSearchQuery(searchQuery);
-
-        return Search
-            .session(entityManager.getEntityManagerFactory().createEntityManager())
-            .search(searchableEntities)
-            .where(function ->
-                function
-                    // https://docs.jboss.org/hibernate/stable/search/reference/en-US/html_single/#search-dsl-predicate-simple-query-string
-                    .simpleQueryString()
-                    .fields(searchableAttributes)
-                    .matching(adaptedSearchQuery)
-                    .defaultOperator(BooleanOperator.AND)
-            )
-            .sort(function -> function.field("lastModifiedDateTime").desc())
-            .fetchAllHits()
-            .stream()
-            .map(searchDomainMapper::entity2Model);
-    }
-
-    protected String createAdaptedSearchQuery(final String searchQuery) {
-        var adaptedSearchQuery = StringUtils.trimToEmpty(searchQuery);
-        final var singleQueryWords = StringUtils.split(adaptedSearchQuery, StringUtils.SPACE);
-        adaptedSearchQuery =
-            Arrays
-                .stream(singleQueryWords)
-                .map(StringUtils::lowerCase)
-                // Anfügen eines Wildcardprefix
-                .map(lowerCaseQueryWord -> lowerCaseQueryWord + "*")
-                // Trennen der Wildcard-Suchwörter mit Leerzeichen
-                .collect(Collectors.joining(StringUtils.SPACE));
-        log.debug("Die erstellte Suchquery: {}", adaptedSearchQuery);
-        return adaptedSearchQuery;
-    }
 
     protected String[] getNamesOfSearchableAttributes(List<Class<? extends BaseEntity>> searchableEntities) {
         final var searchableAttributes = searchableEntities
